Init os3 by hand

+++ os3/src/syscall/process.rs
@@ -15,7 +15,7 @@ pub struct TaskInfo {
     time: usize,
 }
 
-pub fn sys_exit(exit_code: usize) -> ! {
+pub fn sys_exit(exit_code: i32) -> ! {
     info!("[kernel] Application exited with code {}", exit_code);
     exit_current_and_run_next();
     panic!("Unreachable in sys_exit!");
